Migrate GPUProcess structures to generated serialization https://bugs.webkit.org/show_bug.cgi?id=247150 rdar://101651187
Reviewed by Tim Horton.

* Source/WebKit/CMakeLists.txt:
* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/DerivedSources.make:
* Source/WebKit/GPUProcess/GPUProcessSessionParameters.h:
(WebKit::GPUProcessSessionParameters::encode const): Deleted.
(WebKit::GPUProcessSessionParameters::decode): Deleted.
* Source/WebKit/GPUProcess/GPUProcessSessionParameters.serialization.in: Added.
* Source/WebKit/GPUProcess/graphics/RemoteRenderingBackendCreationParameters.h:
(WebKit::RemoteRenderingBackendCreationParameters::encode const): Deleted.
(WebKit::RemoteRenderingBackendCreationParameters::decode): Deleted.
* Source/WebKit/GPUProcess/graphics/RemoteRenderingBackendCreationParameters.serialization.in: Added.
* Source/WebKit/GPUProcess/media/InitializationSegmentInfo.h:
(WebKit::InitializationSegmentInfo::TrackInformation::encode const): Deleted.
(WebKit::InitializationSegmentInfo::TrackInformation::decode): Deleted.
(WebKit::InitializationSegmentInfo::encode const): Deleted.
(WebKit::InitializationSegmentInfo::decode): Deleted.
* Source/WebKit/GPUProcess/media/InitializationSegmentInfo.serialization.in: Added.
* Source/WebKit/GPUProcess/media/MediaDescriptionInfo.h: Copied from Source/WebKit/GPUProcess/graphics/RemoteRenderingBackendCreationParameters.h.
(WebKit::MediaDescriptionInfo::MediaDescriptionInfo):
* Source/WebKit/GPUProcess/media/MediaDescriptionInfo.serialization.in: Added.
* Source/WebKit/GPUProcess/media/RemoteMediaDescription.h:
(WebKit::MediaDescriptionInfo::MediaDescriptionInfo): Deleted.
(WebKit::MediaDescriptionInfo::encode const): Deleted.
(WebKit::MediaDescriptionInfo::decode): Deleted.
* Source/WebKit/GPUProcess/media/TextTrackPrivateRemoteConfiguration.h:
(WebKit::TextTrackPrivateRemoteConfiguration::encode const): Deleted.
(WebKit::TextTrackPrivateRemoteConfiguration::decode): Deleted.
* Source/WebKit/GPUProcess/media/TextTrackPrivateRemoteConfiguration.serialization.in: Added.
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/256084@main
