+++ Source/WebKit/CMakeLists.txt
@@ -516,6 +516,14 @@ endmacro()
 GENERATE_MESSAGE_SOURCES(WebKit_DERIVED_SOURCES "${WebKit_MESSAGES_IN_FILES}")
 
 set(WebKit_SERIALIZATION_IN_FILES
+    GPUProcess/GPUProcessSessionParameters.serialization.in
+
+    GPUProcess/graphics/RemoteRenderingBackendCreationParameters.serialization.in
+
+    GPUProcess/media/InitializationSegmentInfo.serialization.in
+    GPUProcess/media/MediaDescriptionInfo.serialization.in
+    GPUProcess/media/TextTrackPrivateRemoteConfiguration.serialization.in
+
     NetworkProcess/NetworkProcessCreationParameters.serialization.in
 
     Shared/EditorState.serialization.in
