+++ Source/WebKit/GPUProcess/media/TextTrackPrivateRemoteConfiguration.h
@@ -50,118 +50,7 @@ struct TextTrackPrivateRemoteConfiguration {
     bool isMainProgramContent { true };
     bool isEasyToRead { false };
     bool isDefault { false };
-
-    template<class Encoder>
-    void encode(Encoder& encoder) const
-    {
-        encoder << trackId;
-        encoder << label;
-        encoder << language;
-        encoder << inBandMetadataTrackDispatchType;
-        encoder << startTimeVariance;
-        encoder << trackIndex;
-        encoder << cueFormat;
-        encoder << kind;
-        encoder << isClosedCaptions;
-        encoder << isSDH;
-        encoder << containsOnlyForcedSubtitles;
-        encoder << isMainProgramContent;
-        encoder << isEasyToRead;
-        encoder << isDefault;
-    }
-
-    template <class Decoder>
-    static std::optional<TextTrackPrivateRemoteConfiguration> decode(Decoder& decoder)
-    {
-        std::optional<AtomString> trackId;
-        decoder >> trackId;
-        if (!trackId)
-            return std::nullopt;
-
-        std::optional<AtomString> label;
-        decoder >> label;
-        if (!label)
-            return std::nullopt;
-
-        std::optional<AtomString> language;
-        decoder >> language;
-        if (!language)
-            return std::nullopt;
-
-        std::optional<AtomString> inBandMetadataTrackDispatchType;
-        decoder >> inBandMetadataTrackDispatchType;
-        if (!inBandMetadataTrackDispatchType)
-            return std::nullopt;
-
-        std::optional<MediaTime> startTimeVariance;
-        decoder >> startTimeVariance;
-        if (!startTimeVariance)
-            return std::nullopt;
-
-        std::optional<int> trackIndex;
-        decoder >> trackIndex;
-        if (!trackIndex)
-            return std::nullopt;
-
-        std::optional<WebCore::InbandTextTrackPrivate::CueFormat> cueFormat;
-        decoder >> cueFormat;
-        if (!cueFormat)
-            return std::nullopt;
-
-        std::optional<WebCore::InbandTextTrackPrivate::Kind> kind;
-        decoder >> kind;
-        if (!kind)
-            return std::nullopt;
-
-        std::optional<bool> isClosedCaptions;
-        decoder >> isClosedCaptions;
-        if (!isClosedCaptions)
-            return std::nullopt;
-
-        std::optional<bool> isSDH;
-        decoder >> isSDH;
-        if (!isSDH)
-            return std::nullopt;
-
-        std::optional<bool> containsOnlyForcedSubtitles;
-        decoder >> containsOnlyForcedSubtitles;
-        if (!containsOnlyForcedSubtitles)
-            return std::nullopt;
-
-        std::optional<bool> isMainProgramContent;
-        decoder >> isMainProgramContent;
-        if (!isMainProgramContent)
-            return std::nullopt;
-
-        std::optional<bool> isEasyToRead;
-        decoder >> isEasyToRead;
-        if (!isEasyToRead)
-            return std::nullopt;
-
-        std::optional<bool> isDefault;
-        decoder >> isDefault;
-        if (!isDefault)
-            return std::nullopt;
-
-        return {{
-            WTFMove(*trackId),
-            WTFMove(*label),
-            WTFMove(*language),
-            WTFMove(*inBandMetadataTrackDispatchType),
-            WTFMove(*startTimeVariance),
-            *trackIndex,
-            *cueFormat,
-            *kind,
-            *isClosedCaptions,
-            *isSDH,
-            *containsOnlyForcedSubtitles,
-            *isMainProgramContent,
-            *isEasyToRead,
-            *isDefault,
-        }};
-    }
 };
-
 
 } // namespace WebKit
